Dynamic-Query-Visualizer-Llama/requirements.txt
Flask
pandas
requests
sqlalchemy
pyodbc
plotly
langchain
langchain-community
python-dotenv
